Code reorganization

+++ src/Runners/NetworkRunner.cpp
@@ -554,7 +554,7 @@ void NetworkRunner::receiveResults()
 		timer.async_wait(timerHandler);
 		acceptor.async_accept(socket, acceptHandler);
 
-		while (!interrupted && imageParts.size() < imagePartCount)
+		while (!interrupted && (int)imageParts.size() < imagePartCount)
 		{
 			io.run_one();
